Added C-Vibes, tested on tictactoe(10/10 ties). Found a little bug in the MCTS for games which are not TicTacToe(automatically returning +/-(1,) * 8 in Us() func.

+++ CVIBES/PrioritizedItem.py
@@ -3,5 +3,5 @@ from typing import Any
 
 @dataclass(order=True)
 class PrioritizedItem:
-    priority: int
+    priority: float
     item: Any=field(compare=False)
